requirements.txt
Django>=5.2.4
psycopg2-binary>=2.9
gunicorn>=20.1
dj-database-url>=1.0
python-decouple>=3.8
whitenoise==6.5.0  